requirements.txt
SpeechRecognition==3.10.0
pyaudio==0.2.13
flask==2.3.3
flask-cors==4.0.0
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
python-dotenv==1.0.0
groq==0.4.2
pathspec==0.11.2
optree>=0.13.0 